+++ test/endpoints/anime/_root.dart
@@ -1,0 +1,7 @@
+import 'index.dart';
+
+Future<void> anime_test() async {
+  await anime_details();
+  await characters_staff();
+  await episodes();
+}