+++ test/endpoints/anime/index.dart
@@ -1,0 +1,5 @@
+export 'anime_details.dart';
+export 'characters_staff.dart';
+export 'episodes.dart';
+export 'index.dart';
+export '_root.dart';