+++ test/endpoints/manga/_root.dart
@@ -1,0 +1,5 @@
+import 'index.dart';
+
+Future<void> manga_test() async {
+  await manga_details();
+}