+++ test/endpoints/manga/index.dart
@@ -1,0 +1,3 @@
+export 'index.dart';
+export 'manga_details.dart';
+export '_root.dart';